requirements.txt
openai-whisper
openai
mutagen
flask
python-dotenv
requests
pandas
numpy
jsonschema
python-vlc
pygame
matplotlib